requirements.txt
tornado==4.3
Django==1.9.1
django-filter==0.12.0
djangorestframework==3.3.2
Markdown==2.6.5
wheel==0.26.0
codecov==1.6.3
djangorestframework-jwt==1.7.2
httpretty==0.8.14
tmdbsimple==1.3.0